Biofilm for Cystitis
Cystitis, in particular, is the medical term for inflammation of the bladder. Most of the time, the inflammation is caused by a bacterial infection. A bladder infection can be painful and annoying, and it can become a serious health problem if the infection spreads to your kidneys. Typically occurs when bacteria outside the body enter the urinary tract through the urethra and starts to multiply. 

A potential solution to this problem is a biofilm made with a polymeric matrix that has highly adhesive properties that it help to stay in place after hydration. As the oral route may imply an incomplete or diminished absorption I propose the vaginal route. For an effective treatment a constant concentration of the drug in the site is required, and this we achieve with the film. In this way we can avoid different doses of antibiotics.